Created a directory to test scene detect, and set up automated script
Used scene detect successfully, first tested it on howls moving castle, got around 900 frames as compared to 20000 used when analzing the whole film. Then I ran the howls moving castle ipnyb on the reduced frame data and got similar results to the original one so its successfully getting the main colors.

also used a similar script to the one in the Howls moving castle directory and basically updated it to be able run on multiple films